Move compileStylesheet to its own file
